requirements.txt
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-apscheduler==0.5.2
faker==4.18.0
urllib3==1.26.9
requests==2.27.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
pyyaml==5.4.1
pytest-assume==2.4.3
loguru==0.4.1
pytest==7.1.1
pytest-xdist==3.1.0
filelock==3.8.2
jsonpath==0.82
